+++ models/repoHeaderModel.go
@@ -3,7 +3,10 @@ package models
 import (
 	"github.com/charmbracelet/bubbles/v2/paginator"
 	tea "github.com/charmbracelet/bubbletea/v2"
+	"github.com/charmbracelet/lipgloss/v2"
 )
+
+var headerStyle = lipgloss.NewStyle().Foreground(lipgloss.Color("241")).Background(lipgloss.Color("236"))
 
 type TabSelectMessage struct{ Index int }
 
@@ -11,6 +14,7 @@ type RepoHeaderModel struct {
 	titles    []string
 	paginator paginator.Model
 	width     int
+	height    int
 }
 
 func NewRepoHeaderModel(width int, titles []string, index int) RepoHeaderModel {
@@ -27,8 +31,9 @@ func NewRepoHeaderModel(width int, titles []string, index int) RepoHeaderModel {
 	}
 }
 
-func (m *RepoHeaderModel) SetWidth(width int) {
+func (m *RepoHeaderModel) SetDimensions(width, height int) {
 	m.width = width
+	m.height = height
 }
 
 func (m RepoHeaderModel) Init() (tea.Model, tea.Cmd) {
@@ -40,5 +45,6 @@ func (m RepoHeaderModel) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 }
 
 func (m RepoHeaderModel) View() string {
-	return m.paginator.View() + " " + m.titles[m.paginator.Page]
+	heading := headerStyle.Render(m.titles[m.paginator.Page])
+	return heading + "\n" + m.paginator.View()
 }
